# Add ASCII fast path for ILIKE scalar (90% faster)