@tustvold 2022-12-09
I think this might not work for non-ASCII characters that capitalize to ASCII characters :cry: