@alamb 2022-12-09
Perhaps you can also check that left is entirely ASCII (which is a fairly common case)