@tustvold 2022-12-09
Interestingly the regex crate doesn't appear to handle this correctly... :thinking: