@tustvold 2022-12-09
Aargh, there are unicode characters that have lowercase ASCII... Guess I will need to check if the string are ASCII