@tustvold 2022-12-09
This is currently blocked on https://github.com/apache/arrow-rs/pull/3313, which will allow the kernel to assume a GenericStringArray, which in turn will allow ASCII verification in one pass.