@tustvold 2022-12-09
```
ilike_utf8 scalar equals
                        time:   [271.37 µs 271.48 µs 271.62 µs]
                        change: [-87.932% -87.924% -87.914%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 13 outliers among 100 measurements (13.00%)
  2 (2.00%) low mild
  4 (4.00%) high mild
  7 (7.00%) high severe

Benchmarking ilike_utf8 scalar contains: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 9.1s, enable flat sampling, or reduce sample count to 50.
ilike_utf8 scalar contains
                        time:   [1.8028 ms 1.8037 ms 1.8048 ms]
                        change: [-56.517% -56.488% -56.455%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 10 outliers among 100 measurements (10.00%)
  5 (5.00%) high mild
  5 (5.00%) high severe

ilike_utf8 scalar ends with
                        time:   [287.38 µs 287.52 µs 287.65 µs]
                        change: [-87.759% -87.745% -87.729%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high severe

ilike_utf8 scalar starts with
                        time:   [273.73 µs 273.85 µs 273.98 µs]
                        change: [-88.337% -88.328% -88.318%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  1 (1.00%) low mild
  7 (7.00%) high mild
  3 (3.00%) high severe

Benchmarking ilike_utf8 scalar complex: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 9.3s, enable flat sampling, or reduce sample count to 50.
ilike_utf8 scalar complex
                        time:   [1.8338 ms 1.8349 ms 1.8361 ms]
                        change: [-4.1219% -3.9358% -3.7411%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  4 (4.00%) high mild
  5 (5.00%) high severe

nilike_utf8 scalar equals
                        time:   [257.80 µs 257.85 µs 257.92 µs]
                        change: [-89.265% -89.246% -89.211%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  4 (4.00%) high mild
  5 (5.00%) high severe

Benchmarking nilike_utf8 scalar contains: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 9.2s, enable flat sampling, or reduce sample count to 50.
nilike_utf8 scalar contains
                        time:   [1.8283 ms 1.8293 ms 1.8303 ms]
                        change: [-57.048% -56.996% -56.926%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  2 (2.00%) high mild
  4 (4.00%) high severe

nilike_utf8 scalar ends with
                        time:   [256.38 µs 256.46 µs 256.54 µs]
                        change: [-89.212% -89.192% -89.160%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  4 (4.00%) high mild
  4 (4.00%) high severe

nilike_utf8 scalar starts with
                        time:   [262.65 µs 262.70 µs 262.77 µs]
                        change: [-88.974% -88.958% -88.927%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) high mild
  3 (3.00%) high severe

Benchmarking nilike_utf8 scalar complex: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 9.3s, enable flat sampling, or reduce sample count to 50.
nilike_utf8 scalar complex
                        time:   [1.8537 ms 1.8547 ms 1.8559 ms]
                        change: [-3.5514% -3.4042% -3.2212%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 13 outliers among 100 measurements (13.00%)
  2 (2.00%) low mild
  3 (3.00%) high mild
  8 (8.00%) high severe

ilike_utf8_scalar_dyn dictionary[10] string[4])
                        time:   [59.817 µs 59.887 µs 59.968 µs]
                        change: [-97.558% -97.553% -97.546%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 12 outliers among 100 measurements (12.00%)
  5 (5.00%) high mild
  7 (7.00%) high severe
```

So even with the extra check to detect is_ascii, this still is a pretty significant speedup